module structure draft implement
